+++ Runtime/Stores/Android/GooglePlay/AAR/Interfaces/IGoogleBillingClient.cs
@@ -13,7 +13,7 @@ namespace UnityEngine.Purchasing.Interfaces
         GoogleBillingConnectionState GetConnectionState();
         void QueryPurchasesAsync(string skuType, Action<IGoogleBillingResult, IEnumerable<AndroidJavaObject>> onQueryPurchasesResponse);
         void QueryProductDetailsAsync(List<string> skus, string type, Action<IGoogleBillingResult, List<AndroidJavaObject>> onProductDetailsResponseAction);
-        AndroidJavaObject LaunchBillingFlow(AndroidJavaObject productDetails, string oldPurchaseToken, GooglePlayProrationMode? prorationMode);
+        AndroidJavaObject LaunchBillingFlow(AndroidJavaObject productDetails, string oldPurchaseToken, GooglePlayReplacementMode? replacementMode);
         void ConsumeAsync(string purchaseToken, Action<IGoogleBillingResult> onConsume);
         void AcknowledgePurchase(string purchaseToken, Action<IGoogleBillingResult> onAcknowledge);
         void SetObfuscationAccountId(string obfuscationAccountId);