+++ Runtime/Stores/Android/GooglePlay/AAR/Interfaces/IGooglePurchaseService.cs
@@ -2,6 +2,6 @@ namespace UnityEngine.Purchasing.Interfaces
 {
     interface IGooglePurchaseService
     {
-        void Purchase(ProductDefinition product, Product oldProduct, GooglePlayProrationMode? desiredProrationMode);
+        void Purchase(ProductDefinition product, Product oldProduct, GooglePlayReplacementMode? desiredReplacementMode);
     }
 }