+++ Runtime/Stores/Android/GooglePlay/AAR/MetricizedGooglePlayStoreService.cs
@@ -62,10 +62,10 @@ namespace UnityEngine.Purchasing
         }
 
         public override void Purchase(ProductDefinition product, Product? oldProduct,
-            GooglePlayProrationMode? desiredProrationMode)
+            GooglePlayReplacementMode? desiredReplacementMode)
         {
             m_TelemetryMetricsService.ExecuteTimedAction(
-                () => base.Purchase(product, oldProduct, desiredProrationMode),
+                () => base.Purchase(product, oldProduct, desiredReplacementMode),
                 TelemetryMetricDefinitions.initPurchaseName);
         }
     }